app/requirements.txt
openai
flask
flask[async]
flask-wtf>=1.2.0
azure-keyvault-secrets
azure-identity
azure-monitor-opentelemetry
Flask-Session2
werkzeug>=2
requests>=2,<3
identity>=0.5.1,<0.6
msal>=1.26.0
redis
redis-entraid>=0.2.0
python-dotenv